loading board's pages

+++ app/src/main/java/com/mvvm/kanban_board/data/apiService/response/PageResponse.kt
@@ -1,4 +1,4 @@
 package com.mvvm.kanban_board.data.apiService.response
 
-class PageResponse(var id: Long, var name: String, var board: Int) {
+class PageResponse(var id: Long, var name: String, var board: Long) {
 }
